Adds fps delay + test

+++ Ext/misc.c
@@ -2,7 +2,7 @@
 #include <SDL2/SDL.h>
 #include <SDL2/SDL_timer.h>
 
-void CDY_Pause(int time) {
+void CDY_Pause(float time) {
     SDL_Delay(time);
 }
